Soffie - BE, FE, Fx Resto

+++ backend/src/Service/Resto/resto-menu-photos/resto-menu-photos.service.ts
@@ -2,7 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { RestoMenuPhotos } from 'src/entities/RestoMenuPhotos';
 import { Repository } from 'typeorm';
-
+// delete picture
+import { unlink } from 'fs/promises';
 @Injectable()
 export class RestoMenuPhotosService {
     constructor(@InjectRepository(RestoMenuPhotos) private restoMenuPhotoRepository:Repository<RestoMenuPhotos>){}
@@ -33,6 +34,42 @@ export class RestoMenuPhotosService {
             }
         )
     }
+ 
+    async addMultiplePhoto(files:any, body:any){ 
+        const isPrimary = await this.restoMenuPhotoRepository.query(
+            'SELECT * FROM resto.isPrimary($1)',[body.remeId]
+        )
+        const getcount = isPrimary[0].isprimary 
+        
+        for(let i=0; i<files.length; i++){
+            if(getcount === 0 && i === 0){
+                await this.restoMenuPhotoRepository.insert(
+                    {
+                        rempThumbnailFilename: body.rempThumbnailFilename,
+                        rempPhotoFilename: files[i].filename,
+                        rempPrimary: '1',
+                        rempUrl: files[i].path,
+                        rempReme: body.remeId
+        
+                    }
+                )
+            }else{
+                await this.restoMenuPhotoRepository.insert(
+                    {
+                        rempThumbnailFilename: body.rempThumbnailFilename,
+                        rempPhotoFilename: files[i].filename,
+                        rempPrimary: '0',
+                        rempUrl: files[i].path,
+                        rempReme: body.remeId 
+                    }
+                )
+            }
+            
+        } 
+        
+        return 'add successfully' 
+        
+    }
 
             
     async editPrimary(data:any){
@@ -61,7 +98,7 @@ export class RestoMenuPhotosService {
                 // rempThumbnailFilename: body.rempThumbnailFilename,
                 rempPhotoFilename: file.filename,
                 rempPrimary: body.rempPrimary,
-                rempUrl: file.path ,
+                rempUrl: file.path,
                 // rempReme: body.remeId
             }
         )
@@ -69,6 +106,19 @@ export class RestoMenuPhotosService {
 
 
     async deleteMenuPhoto(param){
-        return await this.restoMenuPhotoRepository.delete(param.id)
-    }
+        const id = Number(param.id);
+        // ambil satu data photo untuk di hapus
+        const img = await this.restoMenuPhotoRepository.findOne({
+            where : {
+                rempId:id
+            }
+        });
+        // kalau datanya ada, hapus photonya
+        if(img){
+            // hapus photo di folder restomenuphotos
+            await unlink(img.rempUrl);
+            // hapus satu baris row photo (data photo) di db
+            await this.restoMenuPhotoRepository.delete(param.id)
+        }
+    } 
 }
